feat(templates.neovim): give the neovim module a better template
demo more of the features, make it follow the style of kickstart-nvim
but with lze and nix, make it lazily loaded.
