+++ templates/neovim/flake.nix
@@ -1,10 +1,16 @@
 {
-  description = "Flake exporting a configured package using wlib.evalModule";
+  description = "Flake exporting a configured neovim package";
   inputs.nixpkgs.url = "github:NixOS/nixpkgs/nixpkgs-unstable";
   inputs.wrappers.url = "github:BirdeeHub/nix-wrapper-modules";
   inputs.wrappers.inputs.nixpkgs.follows = "nixpkgs";
-  inputs.nvim-treesitter-textobjects = {
-    url = "github:nvim-treesitter/nvim-treesitter-textobjects/main";
+  # Demo on fetching plugins from outside nixpkgs
+  inputs.plugins-lze = {
+    url = "github:BirdeeHub/lze";
+    flake = false;
+  };
+  # These 2 are already in nixpkgs, however this ensures you always fetch the most up to date version!
+  inputs.plugins-lzextras = {
+    url = "github:BirdeeHub/lzextras";
     flake = false;
   };
   outputs =
