--- templates/neovim/ftplugin/README.md
@@ -1,3 +1,0 @@
-These files are ran on the filetype corresponding to their name.
-
-see [`:h 'rtp'`](https://neovim.io/doc/user/options.html#'rtp')
