--- templates/neovim/plugin/README.md
@@ -1,3 +1,0 @@
-all files in this directory get ran at startup
-
-see [`:h 'rtp'`](https://neovim.io/doc/user/options.html#'rtp')
